requirements.txt
absl-py==2.0.0
accelerate==0.25.0
addict==2.4.0
aiofiles==23.2.1
aiohttp==3.9.1
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
attrs==23.1.0
bleach==6.0.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
deepspeed==0.12.5
einops==0.4.1
exceptiongroup==1.2.0
fastapi==0.100.1
ffmpy==0.3.1
filelock==3.13.1
flash-attn==2.3.6
fonttools==4.47.0
frozenlist==1.4.1
fsspec==2023.12.2
google-auth==2.25.2
google-auth-oauthlib==1.0.0
gradio==3.39.0
gradio-client==0.7.3
grpcio==1.60.0
h11==0.14.0
hjson==3.1.0
httpcore==1.0.2
httpx==0.26.0
huggingface-hub==0.20.1
idna==3.6
imageio==2.33.1
importlib-metadata==7.0.0
jinja2==3.1.2
joblib==1.3.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
kiwisolver==1.4.5
lazy-loader==0.3
linkify-it-py==2.0.2
lmdb==1.4.1
markdown==3.5.1
markdown-it-py==2.2.0
markdown2==2.4.10
markupsafe==2.1.3
matplotlib==3.8.2
mdit-py-plugins==0.3.3
mdurl==0.1.2
msgpack==1.0.7
multidict==6.0.4
networkx==3.2.1
ninja==1.11.1.1
numpy==1.24.2
oauthlib==3.2.2
openai==0.27.8
opencv-python==4.8.0.74
orjson==3.9.10
packaging==23.2
pandas==2.1.4
peft==0.5.0
pillow==9.4.0
pip==23.3.2
platformdirs==4.1.0
protobuf==4.25.1
psutil==5.9.7
py-cpuinfo==9.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycocoevalcap==1.2
pycocotools==2.0.6
pydantic==2.5.2
pydantic-core==2.14.5
pydub==0.25.1
pynvml==11.5.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-multipart==0.0.6
pytz==2023.3.post1
pyyaml==6.0.1
ray==2.6.1
referencing==0.32.0
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.15.2
rsa==4.9
safetensors==0.4.1
scikit-image==0.22.0
scikit-learn==1.3.2
scipy==1.11.2
semantic-version==2.10.0
sentencepiece==0.1.99
setuptools==68.0.0
setuptools-scm==8.0.4
shortuuid==1.0.11
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
tensorboard==2.13.0
tensorboard-data-server==0.7.2
terminaltables==3.1.10
threadpoolctl==3.2.0
tifffile==2023.12.9
tokenizers==0.13.3
tomli==2.0.1
toolz==0.12.0
torch==1.13.1+cu117
torchvision==0.14.1+cu117
tqdm==4.64.1
transformers==4.28.0.dev0
typing-extensions==4.9.0
tzdata==2023.3
uc-micro-py==1.0.2
urllib3==2.1.0
uvicorn==0.23.2
webencodings==0.5.1
websockets==11.0.3
werkzeug==3.0.1
wheel==0.38.4
yapf==0.40.2
yarl==1.9.4
zipp==3.17.0